separate django from drf tests
