+++ runtests/testapp/tests/test_django.py
@@ -2,22 +2,14 @@ from datetime import datetime
 from unittest import skipUnless
 
 from django.core.exceptions import ImproperlyConfigured
-from django.contrib.auth.models import Group
-from django.contrib.sites.models import Site
-from django.urls import reverse
 
 from actstream.signals import action
 from actstream.registry import register, unregister
 from actstream.models import Action, actor_stream, model_stream
 from actstream.tests.base import render, ActivityBaseTestCase
-from actstream.tests.test_drf import BaseDRFTestCase
 from actstream.settings import USE_JSONFIELD
-from actstream.drf.urls import router
-from actstream.drf.serializers import registered_serializers as serializers
 
-from testapp.models import MyUser, Player, Abstract, Unregistered
-from testapp_nested.models.my_model import NestedModel
-from testapp.drf import GroupSerializer
+from testapp.models import MyUser, Abstract, Unregistered
 
 
 class TestAppTests(ActivityBaseTestCase):
@@ -84,19 +76,3 @@ class TestAppTests(ActivityBaseTestCase):
         self.assertEqual(newaction.data['text'], 'foobar')
         self.assertEqual(newaction.data['tags'], ['sayings'])
         self.assertEqual(newaction.data['more_data'], {'pk': self.user.pk})
-
-
-class DRFTestAppTests(BaseDRFTestCase):
-
-    def test_urls(self):
-        self._check_urls('actions', 'follows', 'groups', 'sites',
-                         'players', 'nested-models', 'my-users')
-
-    def test_serializers(self):
-
-        models = (Group, MyUser, Player, Site, NestedModel)
-        self.assertSetEqual(serializers.keys(), models, domap=False)
-
-        groups = self.get(reverse('group-list'))
-        assert len(groups) == 2
-        self.assertSetEqual(GroupSerializer.Meta.fields, groups[0].keys())
